refactor, add route to SignUp page, add login functionality from backend

+++ lib/Model/Ingredient.dart
@@ -1,6 +1,6 @@
 class Ingredient {
   late String name;
-  late int quantity;
+  late double quantity;
   late String unit;
 
   Ingredient(
